Consolidate ProjectList and ProjectSelector into 1 component
